reorganize layers
