+++ subscription-manager/api/server/webhook.go
@@ -13,6 +13,7 @@ type event struct {
 }
 
 func (app *App) webhookHandler(rw http.ResponseWriter, req *http.Request) {
+	// logrus.Debug("webhook endpoint trigged")
 	payload := event{}
 
 	body, err := ioutil.ReadAll(req.Body)
@@ -26,6 +27,7 @@ func (app *App) webhookHandler(rw http.ResponseWriter, req *http.Request) {
 		logrus.WithError(err).Error("fail in decode body")
 		return
 	}
+	// logrus.Debugf("type: %+2v", payload.Type)
 
 	switch payload.Type {
 	case "com.alloycard.core.entities.transaction.TransactionCreatedEvent":
