+++ subscription-manager/internal/jwt/jwt.go
